requirements.txt
anyio==3.6.1
certifi==2022.6.15
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mysql-connector-python-rf==2.2.2
pydantic==1.9.1
PyJWT==2.4.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
